Lay the ground work for an extended help system
The idea is to provide help for each subcommand and to generate manual
pages from them.

The introduces the command 'ici help' and 'ici help COMMAND'.  The former
does the exact same thing as 'ici --help', while the latter simply outputs
the file {COMMAND}.d/help (or /usr/share/ici/{COMMAND}.d/help).  The
contents of these 'help' files should be exactly the form of output that
can be used by help2man to generate manual pages.

Additionally, introduce 'ici --list-cmds', which lists all available
commands.  For each COMMAND, also output the first line of the file
{COMMAND}.d/desc.  These lines should be at most 60 characters long
to fit on the same line as the command itself and some spacing.

Finally, this also adds util/ici.conf, which has ICI_LIB_DIR set to the
current directory (.).  This can be useful to test commands within the
workspace, like this:

	ICI_CONF_DIR=util ./ici --list-cmds